+++ teamscale-gradle-plugin/src/main/kotlin/com/teamscale/TeamscalePlugin.kt
@@ -1,29 +1,27 @@
 package com.teamscale
 
 import com.teamscale.aggregation.TestSuiteCompatibilityUtil
-import com.teamscale.config.extension.TeamscalePluginExtension
-import com.teamscale.config.extension.TeamscaleTestImpactedTaskExtension
+import com.teamscale.extension.TeamscalePluginExtension
+import com.teamscale.extension.TeamscaleTaskExtension
 import com.teamscale.utils.AgentPortGenerator
 import com.teamscale.utils.BuildVersion
+import com.teamscale.utils.jacoco
+import com.teamscale.utils.testing
 import org.gradle.api.GradleException
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.artifacts.Configuration
-import org.gradle.api.model.ObjectFactory
 import org.gradle.api.plugins.JavaPlugin
 import org.gradle.api.plugins.JvmTestSuitePlugin
 import org.gradle.api.plugins.ReportingBasePlugin
 import org.gradle.api.plugins.jvm.JvmTestSuite
 import org.gradle.api.provider.Provider
+import org.gradle.api.tasks.testing.Test
 import org.gradle.kotlin.dsl.create
 import org.gradle.kotlin.dsl.dependencies
-import org.gradle.kotlin.dsl.getByType
 import org.gradle.kotlin.dsl.withType
-import org.gradle.testing.base.TestingExtension
 import org.gradle.testing.jacoco.plugins.JacocoPlugin
-import org.gradle.testing.jacoco.plugins.JacocoTaskExtension
 import org.gradle.util.GradleVersion
-import javax.inject.Inject
 
 
 /**
@@ -64,10 +62,9 @@ abstract class TeamscalePlugin : Plugin<Project> {
 			"*.FastClassByGuice.*",
 			"*.ConstructorAccess"
 		)
-	}
 
-	@get:Inject
-	protected abstract val objectFactory: ObjectFactory
+		val MINIMUM_SUPPORTED_VERSION = GradleVersion.version("8.10")
+	}
 
 	/** The version of the teamscale gradle plugin and impacted-tests-executor.  */
 	private val pluginVersion = BuildVersion.pluginVersion
@@ -78,8 +75,8 @@ abstract class TeamscalePlugin : Plugin<Project> {
 
 	/** Applies the teamscale plugin against the given project.  */
 	override fun apply(project: Project) {
-		if (GradleVersion.current() < GradleVersion.version("8.4")) {
-			throw GradleException("The teamscale plugin requires Gradle version 8.4.0 or higher. Version detected: ${GradleVersion.current()}")
+		if (GradleVersion.current() < MINIMUM_SUPPORTED_VERSION) {
+			throw GradleException("The teamscale plugin requires Gradle version ${MINIMUM_SUPPORTED_VERSION.version} or higher. Version detected: ${GradleVersion.current()}")
 		}
 
 		project.logger.info("Applying teamscale plugin $pluginVersion to ${project.name}")
@@ -93,11 +90,11 @@ abstract class TeamscalePlugin : Plugin<Project> {
 		val pluginExtension =
 			project.extensions.create(TEAMSCALE_EXTENSION_NAME, TeamscalePluginExtension::class.java)
 
-		// Add impacted tests executor to a custom configuration that will later be used to
-		// create the classpath for the TestImpacted created by this plugin.
+		// Add impacted tests executor to a custom configuration that will later be appended to the test classpath
+		// if testwise coverage collection is enabled
 		impactedTestEngineConfiguration = project.configurations.create(IMPACTED_TEST_ENGINE_CONFIGURATION_NAME)
 
-		// Add teamscale jacoco agent to a custom configuration that will later be used to
+		// Add teamscale jacoco agent to a custom configuration that will later be used
 		// to generate testwise coverage if enabled.
 		teamscaleJacocoAgentConfiguration = project.configurations.create(TEAMSCALE_JACOCO_AGENT_CONFIGURATION_NAME)
 
@@ -106,28 +103,35 @@ abstract class TeamscalePlugin : Plugin<Project> {
 			teamscaleJacocoAgentConfiguration("com.teamscale:teamscale-jacoco-agent:$pluginVersion")
 		}
 
-		project.tasks.withType<TeamscaleUpload> {
-			serverConfiguration.convention(pluginExtension.server)
-			repository.convention(pluginExtension.repository)
-			commitDescriptorOrRevision.convention(pluginExtension.commit.combine())
-			ignoreFailures.convention(false)
-			message.convention(partition.map { "$it Gradle upload" })
-		}
+		configureTestTasks(project, pluginExtension)
+		configureTeamscaleUploadTasks(project, pluginExtension)
 
-		// Add the teamscale extension also to all TestImpacted tasks
-		extendTestImpactedTasks(project, pluginExtension)
-
-		// Auto-expose JUnit reports for test tasks bound to JVM test suites
-		val testing = project.extensions.getByType<TestingExtension>()
-		testing.suites.withType<JvmTestSuite> {
+		// Auto-expose JUnit and Testwise coverage reports for test tasks bound to JVM test suites
+		project.testing.suites.withType<JvmTestSuite> {
 			val suite = this
 			suite.targets.configureEach {
-				TestSuiteCompatibilityUtil.exposeJUnitReportsForAggregation(testTask.get(), suite.name)
+				TestSuiteCompatibilityUtil.exposeTestReportArtifactsForAggregation(project, testTask, suite.name)
+			}
+			project.configurations.named(suite.sources.runtimeOnlyConfigurationName) {
+				extendsFrom(impactedTestEngineConfiguration)
 			}
 		}
 	}
 
-	private fun extendTestImpactedTasks(
+	private fun configureTeamscaleUploadTasks(
+		project: Project,
+		pluginExtension: TeamscalePluginExtension
+	) {
+		project.tasks.withType<TeamscaleUpload> {
+			serverConfiguration.convention(pluginExtension.server)
+			repository.convention(pluginExtension.repository)
+			commitDescriptorOrRevision.convention(pluginExtension.commit.combined)
+			ignoreFailures.convention(false)
+			message.convention(partition.map { "$it Gradle upload" })
+		}
+	}
+
+	private fun configureTestTasks(
 		project: Project,
 		teamscalePluginExtension: TeamscalePluginExtension
 	) {
@@ -135,32 +139,26 @@ abstract class TeamscalePlugin : Plugin<Project> {
 			"agent-port-generator",
 			AgentPortGenerator::class.java
 		) {}
-		project.tasks.withType<TestImpacted> {
-			val jacocoTaskExtension = this.extensions.getByType<JacocoTaskExtension>()
-			jacocoTaskExtension.excludes?.addAll(DEFAULT_EXCLUDES)
+		project.tasks.withType<Test> {
+			jacoco.excludes?.addAll(DEFAULT_EXCLUDES) //TODO still needed?
 
-			val extension = this.extensions.create<TeamscaleTestImpactedTaskExtension>(
-				TEAMSCALE_EXTENSION_NAME,
-				objectFactory,
-				teamscaleJacocoAgentConfiguration,
-				jacocoTaskExtension
-			)
-			val port = agentPortGenerator.get().getNextPort()
-			extension.agent.useLocalAgent("http://127.0.0.1:${port}/")
-			extension.agent.destination.set(project.layout.buildDirectory.dir("jacoco/${this.name}"))
+			val extension = this.extensions.create<TeamscaleTaskExtension>(
+				TEAMSCALE_EXTENSION_NAME, teamscaleJacocoAgentConfiguration, jacoco
+			).apply {
+				collectTestwiseCoverage.convention(false)
+				runImpacted.convention(false)
+				runAllTests.convention(false)
+				includeAddedTests.convention(true)
+				includeFailedAndSkipped.convention(true)
+				val port = agentPortGenerator.get().getNextPort()
+				agent.useLocalAgent("http://127.0.0.1:${port}/")
+				agent.destination.set(project.layout.buildDirectory.dir("jacoco/${this@withType.name}"))
+			}
 
-			collectTestwiseCoverage.convention(true)
-			runImpacted.convention(false)
-			runAllTests.convention(false)
-			includeAddedTests.convention(true)
-			includeFailedAndSkipped.convention(true)
-			agentConfiguration.convention(extension.agent)
-			serverConfiguration.convention(teamscalePluginExtension.server)
-			testEngineConfiguration.from(impactedTestEngineConfiguration)
-			endCommit.convention(teamscalePluginExtension.commit.combine())
-			baseline.convention(teamscalePluginExtension.baseline)
-			baselineRevision.convention(teamscalePluginExtension.baselineRevision)
-			repository.convention(teamscalePluginExtension.repository)
+			doFirst("testImpactConfiguration", TestImpactConfigurationAction(teamscalePluginExtension, extension))
+
+			outputs.doNotCacheIf("When using Test Impact Analysis ") { extension.runImpacted.get() }
+			outputs.dir(extension.agent.destination)
 		}
 	}
 }