+++ teamscale-gradle-plugin/src/main/kotlin/com/teamscale/aggregation/testwise/TestwiseCoverageAggregationPlugin.kt
@@ -4,6 +4,7 @@ import com.teamscale.TeamscalePlugin
 import com.teamscale.aggregation.ReportAggregationPlugin
 import com.teamscale.aggregation.testwise.internal.DefaultAggregateTestwiseCoverageReport
 import com.teamscale.utils.PartialData
+import com.teamscale.utils.reporting
 import com.teamscale.utils.testwiseCoverageResults
 import org.gradle.api.Plugin
 import org.gradle.api.Project
@@ -11,7 +12,6 @@ import org.gradle.api.artifacts.component.ProjectComponentIdentifier
 import org.gradle.api.artifacts.type.ArtifactTypeDefinition
 import org.gradle.api.attributes.LibraryElements
 import org.gradle.api.model.ObjectFactory
-import org.gradle.api.reporting.ReportingExtension
 import org.gradle.kotlin.dsl.named
 import org.gradle.kotlin.dsl.withType
 import javax.inject.Inject
@@ -34,15 +34,14 @@ abstract class TestwiseCoverageAggregationPlugin : Plugin<Project> {
 	override fun apply(project: Project) {
 		project.plugins.apply(TeamscalePlugin::class.java)
 
-		val configurations = project.configurations
-
-		val reporting = project.extensions.getByType(ReportingExtension::class.java)
+		val reporting = project.reporting
 		reporting.reports.registerBinding(
 			AggregateTestwiseCoverageReport::class.java,
 			DefaultAggregateTestwiseCoverageReport::class.java
 		)
 
-		val codeCoverageResultsConf = configurations.getByName(ReportAggregationPlugin.RESOLVABLE_REPORT_AGGREGATION_CONFIGURATION_NAME)
+		val codeCoverageResultsConf =
+			project.configurations.getByName(ReportAggregationPlugin.RESOLVABLE_REPORT_AGGREGATION_CONFIGURATION_NAME)
 
 		val classDirectories = codeCoverageResultsConf.incoming.artifactView {
 			componentFilter { it is ProjectComponentIdentifier }
@@ -80,7 +79,13 @@ abstract class TestwiseCoverageAggregationPlugin : Plugin<Project> {
 							ArtifactTypeDefinition.DIRECTORY_TYPE
 						)
 					}
-				}.artifacts.resolvedArtifacts.map { it.any { artifact -> artifact.variant.attributes.getAttribute(PartialData.PARTIAL_DATA_ATTRIBUTE) ?: false } })
+				}.artifacts.resolvedArtifacts.map {
+					it.any { artifact ->
+						artifact.variant.attributes.getAttribute(
+							PartialData.PARTIAL_DATA_ATTRIBUTE
+						) ?: false
+					}
+				})
 			}
 		}
 	}