+++ teamscale-gradle-plugin/src/main/kotlin/com/teamscale/config/Commit.kt
@@ -11,12 +11,12 @@ import javax.inject.Inject
 
 /** The commit object which holds the end commit for which we do Test Impact Analysis. */
 @Suppress("MemberVisibilityCanBePrivate")
-abstract class Commit @Inject constructor(objects: ObjectFactory, val providers: ProviderFactory) : Serializable {
+abstract class Commit @Inject constructor(objects: ObjectFactory, private val providers: ProviderFactory) : Serializable {
 
 	/**
 	 * The branch to which the artifacts belong to.
 	 * This field encapsulates the value set in the gradle config.
-	 * Use [combine] to get a revision or branch and timestamp.
+	 * Use [combined] to get a revision or branch and timestamp.
 	 * It falls back to retrieving the values from the git repository, if not given manually.
 	 */
 	@Deprecated("Use the revision instead")
@@ -25,7 +25,7 @@ abstract class Commit @Inject constructor(objects: ObjectFactory, val providers:
 	/**
 	 * The timestamp of the commit that has been used to generate the artifacts.
 	 * This field encapsulates the value set in the gradle config.
-	 * Use [combine] to get a revision or branch and timestamp.
+	 * Use [combined] to get a revision or branch and timestamp.
 	 * It falls back to retrieving the values from the git repository, if not given manually.
 	 */
 	@Deprecated("Use the revision instead")
@@ -35,7 +35,7 @@ abstract class Commit @Inject constructor(objects: ObjectFactory, val providers:
 	 * The revision of the commit that the artifacts should be uploaded to.
 	 * This is e.g. the SHA1 hash of the commit in Git or the revision of the commit in SVN.
 	 * This field encapsulates the value set in the gradle config.
-	 * Use [combine] to get a revision or branch and timestamp.
+	 * Use [combined] to get a revision or branch and timestamp.
 	 * It falls back to retrieving the values from the git repository, if not given manually.
 	 */
 	val revision: Property<String> =
@@ -45,7 +45,7 @@ abstract class Commit @Inject constructor(objects: ObjectFactory, val providers:
 	 * Checks that a branch name and timestamp are set or can be retrieved from the projects git and
 	 * stores them for later use.
 	 */
-	fun combine(): Provider<Pair<CommitDescriptor?, String?>> {
+	internal val combined: Provider<Pair<CommitDescriptor?, String?>> by lazy {
 		val commitProvider: Provider<Pair<CommitDescriptor?, String?>> =
 			providers.zip(branchName, timestamp) { branch, timestamp ->
 				CommitDescriptor(branch, timestamp.toString()) to null as String?
@@ -53,6 +53,6 @@ abstract class Commit @Inject constructor(objects: ObjectFactory, val providers:
 		val revisionProvider: Provider<Pair<CommitDescriptor?, String?>> = revision.map { null to it }
 		// If timestamp and branch are set manually, prefer to use them
 		// otherwise use revision as 2nd option
-		return commitProvider.orElse(revisionProvider)
+		commitProvider.orElse(revisionProvider)
 	}
 }