+++ teamscale-gradle-plugin/src/test/kotlin/com/teamscale/plugin/fixtures/TestRootProject.kt
@@ -93,7 +93,7 @@ tasks.withType(Test).configureEach {
 		buildFile.appendText(
 			"""
 			
-task unitTest(type: com.teamscale.TestImpacted) {
+task unitTest(type: Test) {
 	useJUnitPlatform {
 		excludeTags 'integration'
 	}
@@ -112,7 +112,12 @@ task unitTest(type: com.teamscale.TestImpacted) {
 	}
 	testClassesDirs = testing.suites.test.sources.output.classesDirs
 	classpath = testing.suites.test.sources.runtimeClasspath
-	partition = 'Unit Tests'
+	teamscale {
+		collectTestwiseCoverage = true
+		runImpacted = System.getProperty("impacted") != null
+		runAllTests = System.getProperty("runAllTests") != null
+		partition = 'Unit Tests'
+	}
 	finalizedBy('unitTestReport')
 }
 
@@ -120,7 +125,7 @@ tasks.register('unitTestReport', com.teamscale.reporting.testwise.TestwiseCovera
 	from(tasks.unitTest)
 }
 
-task integrationTest(type: com.teamscale.TestImpacted) {
+task integrationTest(type: Test) {
 	useJUnitPlatform {
 		includeTags 'integration'
 	}
@@ -128,7 +133,10 @@ task integrationTest(type: com.teamscale.TestImpacted) {
 	jacoco.includes = [ 'com.example.project.*' ]
 	testClassesDirs = testing.suites.test.sources.output.classesDirs
 	classpath = testing.suites.test.sources.runtimeClasspath
-	partition = 'Integration Tests'
+	teamscale {
+		collectTestwiseCoverage = true
+		partition = 'Integration Tests'
+	}
 }
 		""".trimIndent()
 		)