+++ CHANGELOG.md
@@ -7,7 +7,7 @@ We use [semantic versioning](http://semver.org/):
 # Next version
 - [feature] _teamscale-gradle-plugin_: Made compatible with Gradle 9.0
 - [breaking] _teamscale-gradle-plugin_: Support for Gradle configuration cache (see `teamscale-gradle-plugin/MIGRATION_35.md` for how to update)
-- [breaking] _teamscale-gradle-plugin_: Bumped up minimum required Gradle version to 8.4
+- [breaking] _teamscale-gradle-plugin_: Bumped up minimum required Gradle version to 8.10
 
 # 34.2.4
 - [fix] Convert tool could fail with `IOException` when providing nested jars